requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
opencv-python>=4.5.0
pillow>=8.0.0
matplotlib>=3.4.0
tqdm>=4.60.0